Fix low-prom peaks not pruned near tile borders.
Some peaks with computed prominence of -1 were not being
pruned, because the -1 value was reserved as a sentinel
for unknown prominence. Using a different sentinel value
fixes the bug.

The -1 prominence values are coming from runoffs whose
neighbor peaks are pruned, and which are 1 foot higher than
their new parent peak after pruning.  It's not clear how
to prevent that from happening.