Fixed wrong coordinate-system for neigbouring tiles incorrect distance if ilp found in last round.
For neigbouring tiles the coordinate-system of the base-tile
with wrong minLat and minLng was used. This was fixed by
providing the fileFormat to the isolation_finder routine
and creating a new coordinate system with it.
another issue is, when the distance is found in the last round,
only the eucledean distance was used to find the ilp.
Fixed this by not updating the inner ring in the last round,
if an ilp was identified.