Compute boundary of Lidar data differently.
The previous approach of using overviews from the VRT could
be extremely slow, because if even a single source raster
didn't have overviews, or had overviews that didn't qualify
for some reason, the entire VRT would have to use full
resolution to compute overviews.  Now instead we compute the
boundary for each source raster individually and then take the
union.